requirements.txt
flask==3.0.0
anthropic==0.7.0
translate==3.6.1
python-dotenv==1.0.0